publish initial major changes
Version two of the site involves transitioning the one-pager onboarding funnel with a single CTA into a multi-page site that acts as a surface for product/community marketing.

The major change here is the addition of another page, /blog, which acts as an index for publishing activity from within the urbit network.

Other changes include actually surfacing our product ecosystem on the front page.